Changed makeapp.sh to not make xnomad.app "run in background"
By using the -B switch with platypus, xnomad.app will not show an icon in your dock or be at all accessible. To quit xnomad, you have to force quit it via activity monitor. By making it a more normal app, it's easier to quit if needed. Much better for new users.